/* PR_START p--sql-update 09 */ Move column pruner classes to separate module.
