+++ metricflow/sql/optimizer/column_pruning/cte_alias_to_cte_node_mapping.py
