+++ tests_metricflow/sql/optimizer/test_cte_column_pruner.py
@@ -15,7 +15,7 @@ from metricflow_semantics.sql.sql_join_type import SqlJoinType
 from metricflow_semantics.sql.sql_table import SqlTable
 from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfiguration
 
-from metricflow.sql.optimizer.column_pruner import SqlColumnPrunerOptimizer
+from metricflow.sql.optimizer.column_pruning.column_pruner import SqlColumnPrunerOptimizer
 from metricflow.sql.render.sql_plan_renderer import DefaultSqlPlanRenderer, SqlPlanRenderer
 from metricflow.sql.sql_cte_node import SqlCteNode
 from metricflow.sql.sql_plan import (
